requirements-Administrator’s MacBook Pro (2).txt
Django==4.2.8
pylti1p3==2.0.0
canvasapi==3.2.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
pycryptodome==3.18.0
cryptography==41.0.7
django-ratelimit==4.1.0
sentry-sdk==1.40.6

# Tracing and Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-django>=0.40b0
opentelemetry-instrumentation-requests>=0.40b0
opentelemetry-instrumentation-sqlite3>=0.40b0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-exporter-jaeger>=1.20.0
arize>=7.0.0